feat: vendor service cleanup
